add vip + fixes

+++ mpi-backend/src/main/java/itmo/mpi/repository/CrewRepository.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface CrewRepository extends JpaRepository<Crew, Long> {
+public interface CrewRepository extends JpaRepository<Crew, Integer> {
 
     @Query(value = "select cr.* from crew cr left join trip_request tr on cr.uid = tr.crew_id and tr.date_start <= " +
             ":dateEnd and tr.date_end >= :dateStart AND tr.status IN ('COMPLETE', 'APPROVED_BY_CREW', 'APPROVED_BY_BOTH') where tr.uid is null;", nativeQuery = true)
